requirements.txt
google-adk==1.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-cloud-logging==3.9.0
google-cloud-aiplatform==1.95.1
mcp>=1.5.0
deprecated
